requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
openai==0.27.0
langchain==0.0.151
linkedin-api==2.0.0a2
pydantic==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
celery==5.2.7
requests==2.26.0
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1